requirements.txt
# Core data processing
pandas>=2.1.4
numpy>=1.20.0
pyarrow>=10.0.0

# Statistical analysis and causal inference
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
pytesseract>=0.3.13

# Development and notebooks
jupyterlab>=4.0.0

# Google Cloud and BigQuery
google-cloud-bigquery>=3.15.0
google-auth>=2.30.0
pandas-gbq>=0.19.0

# Configuration and utilities
python-dotenv>=1.0.1
tqdm>=4.65.0
pyyaml>=6.0
pathlib2>=2.3.0

# API integrations
requests>=2.25.0

# Testing and quality (development dependencies)
pytest>=7.0.0
pytest-cov>=4.0.0
ruff>=0.1.0
black>=23.0.0
mypy>=1.0.0
pre-commit>=3.0.0
nbstripout>=0.6.0
